Finished Activity and Fixing Minor Bug :*
