+++ app/src/main/java/com/example/habitance/data/Repository.kt
@@ -2,7 +2,6 @@ package com.example.habitance.data
 
 import com.google.firebase.Firebase
 import com.google.firebase.firestore.CollectionReference
-import com.google.firebase.firestore.FieldValue
 import com.google.firebase.firestore.firestore
 import kotlinx.coroutines.tasks.await
 
@@ -13,7 +12,7 @@ class Repository {
         return usersCollection.document(userId).collection("activities")
     }
 
-    suspend fun updateActivityProgress(userId: String, activityId: String, progress: List<Int>) {
+    suspend fun updateActivityProgress(userId: String, activityId: String, progress: Map<String, Int>) {
         activityCollection(userId).document(activityId).update("progress", progress).await()
     }
 }
