+++ app/src/main/java/com/example/habitance/ui/screens/activity_list/ActivityList.kt
@@ -1,5 +1,7 @@
-package com.example.habitance.ui.screens.finishedactivity
+package com.example.habitance.ui.screens.activity_list
 
+import CardList
+import android.util.Log
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -17,6 +19,8 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.BasicTextField
@@ -29,9 +33,9 @@ import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.Icon
-import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -41,39 +45,73 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.navigation.NavHostController
-import androidx.navigation.compose.rememberNavController
+import androidx.navigation.NavController
 import com.example.habitance.R
+import com.example.habitance.navbar.BottomBarScreen
 import com.example.habitance.navbar.Screen
-import com.example.habitance.ui.components.CardFinished
+import com.example.habitance.ui.screens.add_activity.Activity
+import com.example.habitance.ui.screens.add_activity.CategoryActivity
+import com.example.habitance.ui.screens.finished_activity.navigateAndPopUp
 import com.example.habitance.ui.theme.BackGround
 import com.example.habitance.ui.theme.BackGround2
+import com.example.habitance.ui.theme.Border
 import com.example.habitance.ui.theme.TextDark
 import com.example.habitance.ui.theme.TextLight
-import com.example.habitance.ui.theme.TextMedium
 import com.example.habitance.ui.theme.fontFamily
+import com.google.firebase.Timestamp
+import com.google.firebase.auth.FirebaseAuth
+import com.google.firebase.firestore.FirebaseFirestore
 
 @Composable
-fun FinishedActivity(navController: NavHostController){
-    val searchQuery = remember { mutableStateOf("") }
+fun ListActivity(navController: NavController) {
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(BackGround)
+            .padding(0.dp, 15.dp)
+    ) {
+        val activities = remember { mutableStateOf<List<Activity>>(emptyList()) }
+        val filteredActivities = remember { mutableStateOf<List<Activity>>(emptyList()) }
+        val searchQuery = remember { mutableStateOf("") }
+        val errorMessage = remember { mutableStateOf<String?>(null) }
+        val selectedCategory = remember { mutableStateOf(CategoryActivity.Baik) }
 
-    Column(modifier = Modifier
-        .fillMaxSize()
-        .background(BackGround)
-        .padding(0.dp, 15.dp)
-    ){
+        LaunchedEffect(Unit) {
+            fetchActivities(
+                onResult = { fetchedActivities ->
+                    activities.value = fetchedActivities.filter {
+                        it.end.seconds >= Timestamp.now().seconds
+                    }
+                    filteredActivities.value = fetchedActivities
+                    filteredActivities.value = activities.value.filter {
+                        it.category == selectedCategory.value
+                    }
+                },
+                onError = { error ->
+                    errorMessage.value = error.message
+                }
+            )
+        }
+
+        LaunchedEffect(selectedCategory.value) {
+            filteredActivities.value = activities.value.filter {
+                it.category == selectedCategory.value
+            }
+        }
         Box(
             modifier = Modifier
                 .fillMaxWidth()
-        ){
+        ) {
             Icon(
                 imageVector = Icons.AutoMirrored.Filled.ArrowBack,
                 contentDescription = "Back icon",
                 tint = TextDark,
                 modifier = Modifier
+                    .clickable {
+                        navController.navigateAndPopUp("home", BottomBarScreen.Activity.route)
+                    }
                     .align(Alignment.CenterStart)
                     .padding(start = 32.dp)
             )
@@ -84,43 +122,42 @@ fun FinishedActivity(navController: NavHostController){
                     .width(61.dp)
                     .height(51.dp)
                     .align(Alignment.Center)
-                    .clickable { navController.navigate(Screen.HomeScreen.route) }
             )
         }
         Spacer(Modifier.size(15.dp))
-        Card (
+        Card(
             colors = CardDefaults.cardColors(BackGround2),
-            border = BorderStroke(1.dp, TextMedium),
+            border = BorderStroke(1.dp, Border),
             modifier = Modifier
                 .fillMaxSize()
                 .padding(10.dp, 0.dp),
-            elevation = CardDefaults.cardElevation(8.dp), // Untuk bayangan card
-            shape = RoundedCornerShape(16.dp) // Bentuk sudut card
+            elevation = CardDefaults.cardElevation(8.dp),
+            shape = RoundedCornerShape(16.dp)
         ) {
             Column(
                 modifier = Modifier.padding(25.dp)
-            ){
+            ) {
                 Text(
-                    text= "FINISHED ACTIVITY",
+                    text = "ACTIVITY LIST",
                     fontSize = 22.sp,
                     color = TextDark,
                     fontFamily = fontFamily,
                     fontWeight = FontWeight(600)
                 )
                 Spacer(modifier = Modifier.size(20.dp))
-                Row (
+                Row(
                     verticalAlignment = Alignment.CenterVertically,
                     horizontalArrangement = Arrangement.SpaceBetween
-                ){
+                ) {
                     Button(
-                        onClick = {navController.navigate(Screen.AddActivityScreen.route)},
+                        onClick = {
+                            navController.navigate(Screen.AddActivityScreen.route)
+                        },
                         modifier = Modifier
-                            .size(42.dp) // Ukuran lingkaran
-                            .clip(CircleShape), // Membuat bentuk lingkaran
-                        colors = ButtonDefaults.buttonColors(
-                            TextDark // Warna latar belakang tombol
-                        ),
-                        contentPadding = PaddingValues(0.dp) // Hilangkan padding default
+                            .size(42.dp)
+                            .clip(CircleShape),
+                        colors = ButtonDefaults.buttonColors(TextDark),
+                        contentPadding = PaddingValues(0.dp)
                     ) {
                         Icon(
                             imageVector = Icons.Default.Add,
@@ -133,16 +170,16 @@ fun FinishedActivity(navController: NavHostController){
                         value = searchQuery.value,
                         onValueChange = { query ->
                             searchQuery.value = query
-//                            filteredActivities.value = activities.value.filter {
-//                                it.name.contains(query, ignoreCase = true)
-//                            }
+                            filteredActivities.value = activities.value.filter {
+                                it.name.contains(query, ignoreCase = true)
+                            }
                         },
                         modifier = Modifier
                             .background(TextLight, shape = RoundedCornerShape(size = 20.dp))
                             .border(width = 1.dp, color = TextDark, shape = RoundedCornerShape(size = 20.dp))
                             .height(42.dp)
                             .fillMaxWidth(),
-                        decorationBox = { innerTextField ->
+                        decorationBox = { innerTextField ->  
                             Row(verticalAlignment = Alignment.CenterVertically) {
                                 Icon(
                                     imageVector = Icons.Default.Search,
@@ -181,21 +218,16 @@ fun FinishedActivity(navController: NavHostController){
                         onClick = {
                             isClickBaik = true
                             isClickBuruk = false
-//                            selectedCategory.value = "Baik"
-//                            filteredActivities.value = activities.value.filter {
-//                                it.category == selectedCategory.value
-//                            }
+                            selectedCategory.value = CategoryActivity.Baik
                         },
                         modifier = Modifier
                             .weight(1f)
                             .clip(RoundedCornerShape(20.dp)),
-                        colors = if(isClickBaik) ButtonDefaults.buttonColors(TextDark) else ButtonDefaults.buttonColors(
-                            TextLight
-                        ),
+                        colors = if(isClickBaik) ButtonDefaults.buttonColors(TextDark) else ButtonDefaults.buttonColors(TextLight),
                         contentPadding = PaddingValues(10.dp)
                     ) {
                         Text(
-                            text = "Baik",
+                            text = CategoryActivity.Baik.name,
                             color = if(isClickBaik) TextLight else TextDark,
                             fontFamily = fontFamily
                         )
@@ -205,21 +237,16 @@ fun FinishedActivity(navController: NavHostController){
                         onClick = {
                             isClickBaik = false
                             isClickBuruk = true
-//                            selectedCategory.value = "Buruk"
-//                            filteredActivities.value = activities.value.filter {
-//                                it.category == selectedCategory.value
-//                            }
+                            selectedCategory.value = CategoryActivity.Buruk
                         },
                         modifier = Modifier
                             .weight(1f)
                             .clip(RoundedCornerShape(20.dp)),
-                        colors = if(isClickBuruk) ButtonDefaults.buttonColors(TextDark) else ButtonDefaults.buttonColors(
-                            TextLight
-                        ),
+                        colors = if(isClickBuruk) ButtonDefaults.buttonColors(TextDark) else ButtonDefaults.buttonColors(TextLight),
                         contentPadding = PaddingValues(10.dp)
                     ) {
                         Text(
-                            text = "Buruk",
+                            text = CategoryActivity.Buruk.name,
                             color = if(isClickBuruk) TextLight else TextDark,
                             fontFamily = fontFamily
                         )
@@ -227,19 +254,54 @@ fun FinishedActivity(navController: NavHostController){
                 }
                 Spacer(Modifier.size(15.dp))
 
-                CardFinished()
-                CardFinished()
-                CardFinished()
-                CardFinished()
-                CardFinished()
-                CardFinished()
+                LazyColumn {
+                    items(filteredActivities.value) { activity ->
+                        Log.d("ActivityList", "Activity: ${activity.name}")
+                        CardList(
+                            activity = activity,
+                            onNavigateToDetail = {
+                                navController.navigate("detailActivity/${activity.id}")
+                            }
+                        )
+                    }
+                }
             }
         }
     }
 }
 
-@Preview
-@Composable
-fun ActivityListPreview(){
-    FinishedActivity(navController = rememberNavController())
+
+
+fun fetchActivities(
+    onResult: (List<Activity>) -> Unit,
+    onError: (Exception) -> Unit
+) {
+    val firestore = FirebaseFirestore.getInstance()
+    val currentUser = FirebaseAuth.getInstance().currentUser
+
+    if (currentUser != null) {
+        firestore.collection("users")
+            .document(currentUser.uid)
+            .collection("activities")
+            .get()
+            .addOnSuccessListener { querySnapshot ->
+                val activities = querySnapshot.documents.mapNotNull { document ->
+                    try {
+                        Log.d("fetchActivities", "Activity: ${document.id}")
+                        document.toObject(Activity::class.java)?.copy(
+                            id = document.id
+                        )
+                    } catch (e: Exception) {
+                        Log.d("fetchActivities", "Error converting document to Activity: ${e.message}")
+                        null
+                    }
+                }
+                onResult(activities) // Return the list of activities
+            }
+            .addOnFailureListener { exception ->
+                onError(exception) // Return the error
+            }
+    } else {
+        onError(Exception("User is not logged in"))
+    }
 }
