--- app/src/main/java/com/example/habitance/ui/screens/activitylist/ActivityListEmpty.kt
@@ -1,11 +1,0 @@
-package com.example.habitance.ui.screens.activitylist
-
-import androidx.compose.runtime.Composable
-import androidx.navigation.NavHostController
-import com.example.habitance.ui.components.ListEmpty
-
-
-@Composable
-fun ActivityListEmpty(navController: NavHostController){
-    ListEmpty("Add Activity", navController = navController)
-}
