+++ app/src/main/java/com/example/habitance/ui/screens/add_activity/Activity.kt
@@ -7,7 +7,7 @@ data class Activity(
     val id : String = "",
     val name: String = "",
     val unit: String = "",
-    val progress: List<Int> = emptyList(),
+    val progress: Map<String, Int> = emptyMap(),
     val target: Int = 0,
     val category: CategoryActivity = CategoryActivity.Baik,
     val periode : String = "",
@@ -19,14 +19,3 @@ enum class CategoryActivity {
     Baik, Buruk
 }
 
-data class ActivityTest(
-    val id : String = "",
-    val name: String = "",
-    val unit: String = "",
-    val progress: Int = 0,
-    val target: Int = 0,
-    val category: CategoryActivity = CategoryActivity.Baik,
-    val periode : String = "",
-    val start: Timestamp = Timestamp.now(),
-    val end: Timestamp = Timestamp.now()
-)
